requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
cfgv==3.3.1
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1.2
django-cors-headers==4.0.0
django-graphene-permissions==0.0.4
exceptiongroup==1.0.1
factory-boy==3.2.1
Faker==15.1.3
filelock==3.8.0
flake8==5.0.4
frozenlist==1.3.1
gql==3.4.0
graphene==3.1.1
graphene-django==3.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
identify==2.5.6
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
multidict==6.0.2
nodeenv==1.7.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
promise==2.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
slugify==0.0.1
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
virtualenv==20.16.5
yarl==1.8.1
